+++ config/config.exs
@@ -1,5 +1,10 @@
 import Config
 
+unless config_env() == :test do
+  config :nx, :default_backend, EXLA.Backend
+  config :exla, client: :rocm
+end
+
 config :ex_astro,
   spice_kernels: [
     # asteroids