+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## v0.1.1 (2024-01-07)
+
+- standardize all functions outputs to be same as in SPICE
+- return :ok/:error tuples
+- new functions
+  - `bodvcd`
+  - `oscelt`
+  - `conics`
+- more documentation
+
 ## v0.1.0 (2024-01-03)
 
 - initial release